Add join validator to users
